+++ Formula/SwagGen.rb
@@ -1,8 +1,8 @@
 class Swaggen < Formula
   desc "Swagger/OpenAPISpec code generator written in Swift"
   homepage "https://github.com/yonaskolb/SwagGen"
-  url "https://github.com/yonaskolb/SwagGen/archive/3.0.2.tar.gz"
-  sha256 "99029d3d4da56e2ffd1006b4ebd1b046660ecd50f233e2a0cc9dddbae328f5ce"
+  url "https://github.com/yonaskolb/SwagGen/archive/4.1.0.tar.gz"
+  sha256 "f12307abab798fb6c9c5c552af1fc30b1761b1b768d9726b0351f2d3f40ee6e4"
   head "https://github.com/yonaskolb/SwagGen.git"
 
   depends_on :xcode