+++ Templates/Swift/Sources/RequestBehaviour.swift
@@ -25,7 +25,7 @@ public protocol RequestBehaviour {
 
 public enum RequestValidationResult {
     case success(URLRequest)
-    case failure(String)
+    case failure(Error)
 }
 
 // Provides empty defaults so that each function becomes optional
@@ -112,7 +112,6 @@ struct RequestBehaviourGroup {
     }
 }
 
-
 //MARK: Type erased Requests and Responses
 
 public typealias AnyResponse = APIResponse<AnyResponseValue>
@@ -126,7 +125,7 @@ public class AnyRequest: APIRequest<AnyResponseValue> {
 
     init<T>(request: APIRequest<T>) {
         requestPath = request.path
-        super.init(service: request.service.asAny(), parameters: request.parameters, headers: request.headers, encodeBody: request.encodeBody)
+        super.init(service: request.service.asAny(), queryParameters: request.queryParameters, formParameters: request.formParameters, headers: request.headers, encodeBody: request.encodeBody)
     }
 }