+++ CHANGELOG.md
@@ -1,5 +1,63 @@
 # Change Log
-## Master
+## Next Version
+
+### Swift Template Updates
+- Fixed path to `Enum.swift` on Linux
+- Added support for objects in query params
+- Added support for nullable properties
+
+## 4.1.0
+
+### Added
+- Added swift template option `enumUndecodableCase` that adds an `undecodable` case to enums when decoding fails #141
+
+### Fixed
+- Fixed installing in Swift 5 #139
+- Fixed Swift template building in Swift 5 by Alamofire #139
+
+### Changed:
+- Updated codebase to Swift 5 and dropped Swift 4.2 #139
+
+[Commits](https://github.com/yonaskolb/SwagGen/compare/4.0.0...4.1.0)
+
+## 4.0.0
+
+### Added
+- Added support for OpenAPISpec/Swagger 3. Support for Swagger 2 has been removed. For that please use release 3.0.2 or the `swagger_2` branch #118 #121
+- Added [StencilSwiftKit](https://github.com/SwiftGen/StencilSwiftKit) support for templates #111
+- Added `oneOf` and `anyOf` with discriminators #121
+- Added support for generating inline schemas when they are wrapped in an array #121
+
+### Swift Template Updates
+- Swagger 3 support #118
+- Added generated `Server` #118
+- Discriminated `oneOf` and `anyOf` enums #121
+- Allow both form and path parameters in the same request #118
+- Add headers to request #120
+- Add framework `Info.plist` #117
+- Use `safeArrayDecoding` #117
+- Catch `APIClientError` from RequestBehaviour validation #117
+- Added `typeAliases` option #117
+- Validation error changed from a String to an Error #117
+- Improve request description and summary #117
+- Change `SecurityRequirement.scope` string to `SecurityRequirement.scopes` array #117
+- Use `StringCodingKey` instead of enum types #117
+- Replace `DateTime` with `Date` #117
+- Update `Alamofire` dependency to 4.8.1 #123
+- Update `Result` dependency to 4.1.0 #123
+- Enums conform to Equatable and CaseIterable #124
+- Removed support for Swift 4.1 #124
+- Only generate `isEqual` in model classes not structs #117
+- Fixed path params that don't have swift friendly names #130
+- Fixed operations with mutiple success responses and no error responses #127
+- Fix nested schemas in subclasses thinking they have a parent #128
+- Handle nil `modelProtocol` option #117
+
+### Removed
+- Removed support for Swagger 2 #118
+- Removed Swift 4.1 support #134
+
+[Commits](https://github.com/yonaskolb/SwagGen/compare/3.0.2...4.0.0)
 
 ## 3.0.2